Add files via upload
This version of the Motor Gimbal varied off course from the other iterations in that we experimented with the idea of using a jig or mount. This mount would be glued inside the rocket tube, and then the gimbal would be screwed into it, eliminating any screw heads from showing outside the body tube. Ultimately, we decided that due to complications that would arise from having objects be unremovable from the body tube, among other reasons, that it would be best/most efficient to just continue with screwing in the gimbal from the outsdie.